requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.5
pyee==7.0.1
requests==2.23.0
six==1.14.0
slackclient==2.5.0
slackeventsapi==2.1.0
urllib3==1.25.9
Werkzeug==1.0.1
yarl==1.4.2
